requirements/discord-bot.txt
discord.py==2.5.2
httpx==0.27.2
python-dotenv==1.0.1
